TESTES - Adição de testes de restaurante e produto
