+++ GastroHub/src/main/java/com/fiap/GastroHub/modules/products/dtos/ProductResponse.java
@@ -6,6 +6,7 @@ import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
+@NoArgsConstructor
 public class ProductResponse {
     private String name;
     private BigDecimal price;
