+++ GastroHub/src/main/java/com/fiap/GastroHub/modules/restaurants/dtos/RestaurantResponse.java
@@ -4,11 +4,12 @@ import com.fiap.GastroHub.modules.users.dtos.UserResponse;
 import lombok.*;
 
 @Data
+@AllArgsConstructor
+@NoArgsConstructor
 public class RestaurantResponse {
     private Long id;
     private String name;
     private String address;
-    private String email;
     private String cuisineType;
     private String openingHours;
     private UserResponse owner;
