+++ GastroHub/src/main/java/com/fiap/GastroHub/modules/restaurants/infra/orm/entities/Restaurant.java
@@ -3,13 +3,19 @@ package com.fiap.GastroHub.modules.restaurants.infra.orm.entities;
 import com.fiap.GastroHub.modules.products.infra.orm.entities.Product;
 import com.fiap.GastroHub.modules.users.infra.orm.entities.User;
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.util.List;
 
 @Data
 @Entity
 @Table(name = "restaurants")
+@Builder(toBuilder = true)
+@NoArgsConstructor
+@AllArgsConstructor
 public class Restaurant {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
