+++ GastroHub/src/main/java/com/fiap/GastroHub/modules/users/dtos/UserResponse.java
@@ -5,6 +5,8 @@ import lombok.*;
 import java.time.LocalDateTime;
 
 @Data
+@AllArgsConstructor
+@NoArgsConstructor
 public class UserResponse {
     private long id;
     private String name;
