+++ GastroHub/src/test/java/com/fiap/GastroHub/modules/restaurants/usecases/GetRestaurantByIdUseCaseTest.java
@@ -1,0 +1,134 @@
+package com.fiap.GastroHub.modules.restaurants.usecases;
+
+import com.fiap.GastroHub.modules.restaurants.dtos.CreateUpdateRestaurantRequest;
+import com.fiap.GastroHub.modules.restaurants.dtos.RestaurantResponse;
+import com.fiap.GastroHub.modules.restaurants.exceptions.RestaurantException;
+import com.fiap.GastroHub.modules.restaurants.infra.orm.entities.Restaurant;
+import com.fiap.GastroHub.modules.restaurants.infra.orm.repositories.RestaurantRepository;
+import com.fiap.GastroHub.modules.roles.infra.orm.entities.Role;
+import com.fiap.GastroHub.modules.users.dtos.UserResponse;
+import com.fiap.GastroHub.modules.users.infra.orm.entities.User;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.MockitoAnnotations;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.modelmapper.ModelMapper;
+import org.springframework.http.HttpStatus;
+
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.Optional;
+
+import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.*;
+
+@ExtendWith(MockitoExtension.class)
+@DisplayName("Get Restaurant By ID Use Case Test Class")
+public class GetRestaurantByIdUseCaseTest {
+    @Mock
+    private RestaurantRepository restaurantRepository;
+
+    private GetRestaurantByIdUseCase getRestaurantByIdUseCase;
+
+    private ModelMapper modelMapper;
+
+    AutoCloseable mock;
+
+    @BeforeEach
+    void setUp() {
+        mock = MockitoAnnotations.openMocks(this);
+        modelMapper = Mockito.spy(new ModelMapper());
+        getRestaurantByIdUseCase = new GetRestaurantByIdUseCase(restaurantRepository, modelMapper);
+    }
+
+    @AfterEach
+    void teardown() throws Exception {
+        mock.close();
+    }
+
+    @Test
+    @DisplayName("Should get a restaurant by it's given id")
+    void execute_ValidId_ReturnsRestaurant() {
+        Role role = Role.builder()
+                .id(1L)
+                .name("Owner")
+                .build();
+
+        User owner = User.builder()
+                .id(1L)
+                .name("User 1")
+                .address("Address 1")
+                .email("email@email.com")
+                .password("password")
+                .role(role)
+                .build();
+
+        UserResponse userResponse = new UserResponse(
+                1L,
+                "Jorge",
+                "jorge@email.com",
+                "Address",
+                LocalDateTime.ofInstant(Instant.now(), ZoneId.systemDefault()),
+                LocalDateTime.ofInstant(Instant.now(), ZoneId.systemDefault())
+        );
+
+        CreateUpdateRestaurantRequest createUpdateRestaurantRequest = new CreateUpdateRestaurantRequest(
+                "Restaurant Name",
+                "Address",
+                "Cuisine Type",
+                "09h00 - 18h00",
+                owner.getId()
+        );
+
+        RestaurantResponse restaurantResponse = new RestaurantResponse(
+                1L,
+                "Restaurant Name",
+                "Address",
+                "Cuisine Type",
+                "09h00 - 18h00",
+                userResponse
+        );
+
+        Restaurant restaurant = Restaurant.builder()
+                .id(1L)
+                .name("Restaurant Name")
+                .address("Restaurant Address")
+                .cuisineType("Restaurant CuisineType")
+                .openingHours("Restaurant OpeningHours")
+                .owner(owner)
+                .build();
+
+        doReturn(restaurantResponse)
+                .when(modelMapper)
+                .map(any(Restaurant.class), eq(RestaurantResponse.class));
+
+        when(restaurantRepository.findById(1L)).thenReturn(Optional.of(restaurant));
+
+        RestaurantResponse result = getRestaurantByIdUseCase.execute(1L);
+
+        assertNotNull(result);
+        assertEquals(restaurant.getName(), result.getName());
+        verify(restaurantRepository, times(1)).findById(1L);
+    }
+
+    @Test
+    @DisplayName("Should not get a restaurant by it's given id")
+    void execute_InvalidId_ThrowsRestaurantException() {
+        when(restaurantRepository.findById(1L)).thenReturn(Optional.empty());
+
+        RestaurantException exception = assertThrows(RestaurantException.class, () -> getRestaurantByIdUseCase.execute(1L));
+
+        assertEquals("Restaurant not found", exception.getMessage());
+        assertEquals(HttpStatus.BAD_REQUEST, exception.getStatusCode());
+        verify(restaurantRepository, times(1)).findById(1L);
+    }
+}
